Linter

+++ contracts/IPendingOrders.sol
@@ -1,7 +1,9 @@
 pragma solidity ^0.7.4;
+
 // "SPDX-License-Identifier: MIT"
 
 interface IPendingOrders {
-    function eventStart(uint _eventId) external;
-    function eventEnd(uint _eventId) external;
+    function eventStart(uint256 _eventId) external;
+
+    function eventEnd(uint256 _eventId) external;
 }
